requirements.txt
amqp==2.6.1
appdirs==1.4.4
attrs==20.1.0
backcall==0.2.0
billiard==3.6.3.0
black==19.10b0
celery==4.4.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
decorator==4.4.2
Django==2.2.6
django-redis==4.12.1
djangorestframework==3.11.0
idna==2.10
importlib-metadata==1.7.0
ipython==7.17.0
ipython-genutils==0.2.0
jedi==0.17.2
kombu==4.6.11
parso==0.7.1
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.6
ptyprocess==0.6.0
Pygments==2.6.1
pytz==2020.1
redis==3.5.2
regex==2020.7.14
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.10
vine==1.3.0
wcwidth==0.2.5
zipp==3.1.0
